# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Go Example REST API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/v1",
    "paths": {
        "/bookmark/append": {
            "post": {
                "description": "Append bookmark to collection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookmark"
                ],
                "summary": "Append bookmark",
                "operationId": "append",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Bookmark"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bookmark/{uuid}": {
            "get": {
                "description": "Show info bookmark",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookmark"
                ],
                "summary": "Show bookmark",
                "operationId": "view",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bookmark UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Bookmark"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ErrorContext": {
            "type": "object",
            "properties": {
                "request_id": {
                    "type": "string"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "context": {
                    "$ref": "#/definitions/handler.ErrorContext"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "model.Bookmark": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "value": {
                    "description": "основное значение, которое нужно запомнить",
                    "type": "string"
                }
            }
        }
    }
}